moved ressource to own file
